Split AST declaration collection helpers
